fixed stuff and also xml layouts mostly
