+++ app/build.gradle
@@ -82,7 +82,7 @@ dependencies {
     // optional - Kotlin Extensions and Coroutines support for Room
     implementation "androidx.room:room-ktx:$room_version"
 
-    implementation 'com.squareup.retrofit2:retrofit:2.8.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.8.1'
 
     implementation 'com.android.support:multidex:1.0.3'
