+++ app/src/main/java/com/example/coronadiagnosticapp/ui/activities/oxymeter/OxymeterThreadEventListener.java
@@ -1,0 +1,13 @@
+package com.example.coronadiagnosticapp.ui.activities.oxymeter;
+
+interface OxymeterThreadEventListener {
+    void onFrame(int frameNumber);
+
+    void onSuccess(Oxymeter oxymeter);
+
+    void onFingerRemoved();
+
+    void onInvalidData();
+
+    void onStartWithNewOxymeter();
+}
