+++ app/src/main/res/layout/camera_fragment.xml
@@ -11,10 +11,12 @@
         android:id="@+id/daily_metrics_message"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="24dp"
+        android:layout_marginEnd="16dp"
         android:gravity="center"
         android:text="@string/header_camera"
         android:textAppearance="@style/appSubHeader"
-        app:layout_constraintBottom_toTopOf="@id/guideline_camera"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -27,23 +29,18 @@
         android:src="@drawable/ic_undraw_doctors_hwty_1"
         app:layout_constraintBottom_toTopOf="@id/button_startCamera"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/daily_metrics_message" />
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="@string/camera_body"
         android:textAppearance="@style/appParagraph"
-        app:layout_constraintBottom_toTopOf="@+id/imageView_background_camera"
+        app:layout_constraintBottom_toBottomOf="@+id/imageView_background_camera"
         app:layout_constraintStart_toStartOf="@+id/imageView_background_camera"
-        app:layout_constraintTop_toTopOf="@+id/guideline_camera" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/guideline_camera"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintGuide_percent="0.3" />
+        app:layout_constraintTop_toBottomOf="@+id/daily_metrics_message"
+        app:layout_constraintVertical_bias="0.36" />
 
     <Button
         android:id="@+id/button_startCamera"
@@ -63,6 +60,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        tools:visibility="visible" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
