+++ app/src/main/res/layout/fragment_question.xml
@@ -15,7 +15,10 @@
         android:layout_height="wrap_content"
         android:fontFamily="sans-serif-condensed"
         android:padding="8dp"
-
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="16dp"
         android:textColor="@android:color/black"
         tools:text="Question title" />
 
@@ -35,6 +38,7 @@
         android:layout_height="wrap_content"
         android:layout_gravity="center"
         android:layout_marginBottom="16dp"
+        android:layout_marginTop="16dp"
         android:text="@string/next" />
 
 </LinearLayout>
