+++ app/src/main/res/layout/fragment_questionnaire.xml
@@ -14,6 +14,9 @@
         android:layout_height="wrap_content"
         android:padding="8dp"
         android:text="Answer some daily questions"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="16dp"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ProgressBar
